--with_mysql=WITH_MYSQL
Set to "yes" if you wish to interface with the MySQL database.
If you choose to have the MySQL interface, you should have the MySQL
client part installed in the machine where you are compiling and using
it.  The MySQL daemon should also be up and running when using the
interface.

Valid values are [yes,no]

--mysql_client_directory=MYSQL_CLIENT_DIRECTORY

You should also specify where the MySQL client library is installed.

git-svn-id: https://babel.ls.fi.upm.es/svn/negation_and_fuzzy_logic@367 f9b316aa-e3e8-4615-9afe-4ac41bce987a